Adding library. Captures of debugging couldn't be saved for some reason. Closes #7
